requirements.txt
absl-py==1.0.0
altair==4.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==5.0.0
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
commonmark==0.9.1
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.6
Eel==0.14.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
extra-streamlit-components==0.1.53
fastjsonschema==2.15.3
filelock==3.0.12
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.4.0
gevent==21.1.2
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.46.1
h11==0.13.0
h5py==3.6.0
idna==3.2
image-quality==1.2.7
imageio==2.19.1
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.0.1
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
libsvm==3.23.0.4
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
numpy==1.21.2
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pipenv==2021.5.29
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyngrok==5.1.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.7.1
selenium==4.1.3
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
stack-data==0.2.0
streamlit==1.10.0
tenacity==8.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.13.3
threadpoolctl==2.2.0
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.6
validators==0.19.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
whichcraft==0.6.1
widgetsnbextension==3.6.0
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
